+++ src/index.js
@@ -1,6 +1,8 @@
 import refs from "./js/refs";
 import API from "./js/api-service";
-import { appendGallery } from "./js/create-markup";
+import { renderFilmList } from "./js/create-markup";
+import { renderMoviesCard } from "./js/create-markup";
+
 
 
 // const newAPI = new API();
