+++ src/js/create-markup.js
@@ -1,23 +1,70 @@
 import refs from "./refs";
+import { addCurrrentMoviesToLocalStorage } from "./local-storage"
 
 
+const IMG_URL = "https://image.tmdb.org/t/p/w500"
+
+///${(release_date || first_air_date) ${film.release_date}
+///////////// --функція для створення карток популярних фільмів на головну сторінку----///////
+function renderFilmList(films) {
+    const markup = films
+        .map((film) => {
+        console.log(film)
+            return `
+    <li class="main-container--card">
+        <img class="film-poster" src="${IMG_URL}${film.poster_path}" alt="${film.original_name}" loading="lazy">
+        <p class="film-info">
+        <h2 class="film-title">${film.original_title.toUpperCase() || film.title.toUpperCase() || film.title.toUpperCase()}</h2>
+        <p class="more-info"> ${transformId(film.genre_ids)} | ${(film.release_date || first_air_date).slice(0,4)}</p>
+        </p>
+    </li>
+  `;
+        }).join("");
+refs.containerBox.innerHTML = markup;
+}
+
+export {renderFilmList};
+
+///////////// --функція для створення карток  фільмів в MY LIBRARY----///////
+function renderMoviesCard(films) {
+    const filmCards = films
+    .map((film) => {
+        return `
+    <li class="main-container--card">
+        <img class="film-poster" src="${IMG_URL}${film.poster_path}" alt="${film.original_name}" loading="lazy">
+        <div class="film-info">
+        <h2 class="film-title">${film.original_title.toUpperCase() || film.title.toUpperCase() || film.title.toUpperCase()}</h2>
+        <p class="more-info"> ${transformId(film.genre_ids)} | ${(film.release_date || first_air_date).slice(0,4)} <span class="vote"> ${film.vote_average} </span> </p>
+        
+        </div>
+    </li>`;
+        }).join("");
+    
+    refs.moviesContainer.insertAdjacentHTML('beforeend', filmCards);
+};
+
+export {renderMoviesCard};
 
 
-//------- НАБРОСОК РЕНДЕР ГАЛЕРЕИ
-
-// const IMG_URL = "https://image.tmdb.org/t/p/w500"
-// function createContent(array) { 
-//     return array.reduce((acc, arr) => acc + createCard(arr), "");
-// }
-
-// function createCard(film) {
-//     return ` <li class="main-container--card"> <img src="${IMG_URL}${film.poster_path}"  alt=""> </li>`;
-// }
-
-// function appendGallery(films) { 
-//     const markup = createContent(films.results);
-//     refs.galleryHome.insertAdjacentHTML("beforeend", markup);
-//     addCurrrentMoviesToLocalStorage(films.results);
-// }
-
-// export {appendGallery};
+///////////// --функція для створення списку жанрів----///////
+function transformId ([...arr]) {
+    const g = localStorage.getItem("GENRES");
+    const genres = JSON.parse(g);
+    let genreName;
+    const array = [...arr]
+    // console.log(genres)
+    for (let i = 0; i < genres.length; i++) {
+        // console.log(genres[i]);
+        for (let x = 0; x < array.length; x++) {
+            if (array[x] === genres[i].id) {
+                genreName = genres[i].name;
+                array[x] = genreName
+            };
+        };
+    };
+    console.log(array)
+    if (array.length > 2) {
+        return `${array[0]}, ${array[1]}, Other`
+    }
+    return `${array[0]}, ${array[1]}`
+};
