+++ src/js/fetch-films-main.js
@@ -1,6 +1,9 @@
 import axios from 'axios'
 import refs from './refs';
 import { addCurrrentMoviesToLocalStorage } from "./local-storage"
+import { renderFilmList } from "./create-markup"
+
+
 
 const API_KEY = '2994e3a31c3cad99fd99bf3fe61d916f';
 const IMG_URL = "https://image.tmdb.org/t/p/w500";
@@ -28,44 +31,46 @@ const addGenresToLocalStorage = async () => {
 };
 
 addGenresToLocalStorage();
-
-function transformId ([...arr]) {
-    const g = localStorage.getItem("GENRES");
-    const genres = JSON.parse(g);
-    let genreName;
-    const array = [...arr]
-    // console.log(genres)
-    for (let i = 0; i < genres.length; i++) {
-        // console.log(genres[i]);
-        for (let x = 0; x < array.length; x++) {
-            if (array[x] === genres[i].id) {
-                genreName = genres[i].name;
-                array[x] = genreName
-            };
-        };
-    };
-    console.log(array)
-    if (array.length > 2) {
-        return `${array[0]}, ${array[1]}, Other`
-    }
-    return `${array[0]}, ${array[1]}`
-};
+///////////// --функція перенесена в файл create-markup----///////
+// function transformId ([...arr]) {
+//     const g = localStorage.getItem("GENRES");
+//     const genres = JSON.parse(g);
+//     let genreName;
+//     const array = [...arr]
+//     // console.log(genres)
+//     for (let i = 0; i < genres.length; i++) {
+//         // console.log(genres[i]);
+//         for (let x = 0; x < array.length; x++) {
+//             if (array[x] === genres[i].id) {
+//                 genreName = genres[i].name;
+//                 array[x] = genreName
+//             };
+//         };
+//     };
+//     console.log(array)
+//     if (array.length > 2) {
+//         return `${array[0]}, ${array[1]}, Other`
+//     }
+//     return `${array[0]}, ${array[1]}`
+// };
 // transformId ([18, 90, 28])
 
-function renderFilmList(films) {
-        const markup = films
-            .map((film) => {
-            console.log(film)
-                return `
-        <li class="main-container--card">
-            <img class="film-poster" src="${IMG_URL}${film.poster_path}" alt="${film.original_title}" loading="lazy">
-            <p class="film-info">
-            <h2 class="film-title">${film.original_title.toUpperCase()}</h2>
-            <p class="more-info"> ${transformId(film.genre_ids)} | ${film.release_date.slice(0,4)}</p>
-            </p>
-        </li>
-      `;
-            }).join("");
-    refs.containerBox.innerHTML = markup;
-};
+
+///////////// --функція перенесена в файл create-markup----///////
+// function renderFilmList(films) {
+//         const markup = films
+//             .map((film) => {
+//             console.log(film)
+//                 return `
+//         <li class="main-container--card">
+//             <img class="film-poster" src="${IMG_URL}${film.poster_path}" alt="${film.original_title}" loading="lazy">
+//             <p class="film-info">
+//             <h2 class="film-title">${film.original_title.toUpperCase()}</h2>
+//             <p class="more-info"> ${transformId(film.genre_ids)} | ${film.release_date.slice(0,4)}</p>
+//             </p>
+//         </li>
+//       `;
+//             }).join("");
+//     refs.containerBox.innerHTML = markup;
+// };
