+++ src/js/refs.js
@@ -3,6 +3,7 @@ export default  {
     containerBox: document.querySelector(`.main-container--card__box`),
     containerCard: document.querySelector(`.main-container--card`),
     watchedList: document.querySelector(".header_btn-watched"),
+    moviesContainer: document.querySelector('.main-container--card__box'),
     queueList: document.querySelector(".header_btn-queue")
     // btnAddWatched: document.querySelector(".watched-btn"),
     // btnQueue: document.querySelector(".queue-btn")
