+++ src/js/search-by-keyword.js
@@ -1,4 +1,5 @@
 import axios from "axios";
+import { renderMoviesCard } from "./create-markup"
 
 const refs = {
     searchForm: document.querySelector('.header_input-wrap'),
@@ -50,51 +51,51 @@ async function onSearch(event) {
     refs.inputEl.value = "";
 }
 
-//Наступна функція буде винесена в окремий файл
-function renderMoviesCard(results) {
-    const markup = results
-        .map(el => {
-            return `
-            <li class="main-container--card">
-            <img src = "https://image.tmdb.org/t/p/w500${el.poster_path}"
-            class = "film-poster"
-            alt="${el.original_title}" loading="lazy"
-             />
-            <div class="film-info">
-            <h3 class="film-title">${el.original_title.toUpperCase()}</h3>
-            <p class="more-info">${transformId(el.genre_ids)} | ${el.release_date.slice(0, 4)}</p>
-            </div>
-            </li> `
-        })
-        .join('');
+///////////// --функція перенесена в файл create-markup----///////
+// function renderMoviesCard(results) {
+//     const markup = results
+//         .map(el => {
+//             return `
+//             <li class="main-container--card">
+//             <img src = "https://image.tmdb.org/t/p/w500${el.poster_path}"
+//             class = "film-poster"
+//             alt="${el.original_title}" loading="lazy"
+//              />
+//             <div class="film-info">
+//             <h3 class="film-title">${el.original_title.toUpperCase()}</h3>
+//             <p class="more-info">${transformId(el.genre_ids)} | ${el.release_date.slice(0, 4)}</p>
+//             </div>
+//             </li> `
+//         })
+//         .join('');
     
-    refs.moviesContainer.insertAdjacentHTML('beforeend', markup);
-};
+//     refs.moviesContainer.insertAdjacentHTML('beforeend', markup);
+// };
 
-//Функція нижче також буде винесена в інший файл 
-function transformId ([...arr]) {
-    const g = localStorage.getItem("GENRES");
-    const genres = JSON.parse(g);
-    let genreName;
-    const array = [...arr]
-    for (let i = 0; i < genres.length; i++) {
-        for (let y = 0; y < array.length; y++) {
-            if (array[y] === genres[i].id) {
-                genreName = genres[i].name;
-                array[y] = genreName
-            };
-        };
-    };
-    // console.log(array)
-    if (array.length > 3) {
-        return `${array[0]}, ${array[1]}, Other`
-    } else if (array.length > 2) {
-        return `${array[0]}, ${array[1]}, ${array[2]}`
-    } else if (array.length > 1) {
-        return `${array[0]}, ${array[1]}`
-    }
-    return `${array[0]}`
-};
+///////////// --функція перенесена в файл create-markup----///////
+// function transformId ([...arr]) {
+//     const g = localStorage.getItem("GENRES");
+//     const genres = JSON.parse(g);
+//     let genreName;
+//     const array = [...arr]
+//     for (let i = 0; i < genres.length; i++) {
+//         for (let y = 0; y < array.length; y++) {
+//             if (array[y] === genres[i].id) {
+//                 genreName = genres[i].name;
+//                 array[y] = genreName
+//             };
+//         };
+//     };
+//     // console.log(array)
+//     if (array.length > 3) {
+//         return `${array[0]}, ${array[1]}, Other`
+//     } else if (array.length > 2) {
+//         return `${array[0]}, ${array[1]}, ${array[2]}`
+//     } else if (array.length > 1) {
+//         return `${array[0]}, ${array[1]}`
+//     }
+//     return `${array[0]}`
+// };
 
 
 function clearMoviesContainer() {
