+++ src/sass/_film_container_main.scss
@@ -109,3 +109,16 @@
        margin: 32px 16px 0 0;
    }
    }
+
+
+   .vote {
+    display: inline-block;
+    text-align: center;
+    width: 36px;
+    height: 15px;
+    text-transform: uppercase;
+    color: #fff;
+    background: #ff6b08;
+    border-radius: 5px;
+    border: 1px solid #ff6b08;
+   }
